Create Microcontroller.vhd
Since I've sent you the old file, there are a couple of updates:

1. There is now a signal called "address_sel" which is used for anything involving memory operations.
2. I've added a "Conv_to_Int", which is a function that converts a std_logic_vector to an integer (useful in conjunction with address_sel).
3. The clock cycle execution and the reset are now in the same process (previously, VHDL was complaining about a conflict between the two processes during compilation).

If you've already done some work in this area, that's fine. However, after looking into it some more, I don't believe we can use components inside a process, which means we have to use functions, or else just add the instructions in right after the opcode.